make version 6.x compatible with php8.0
